Feat: url 변경 및 컴포넌트 적용
